Fixed clean docker script

+++ commands/developer-utils/docker/clean-docker.sh
@@ -1,8 +1,10 @@
 #!/bin/bash
 
+ # Dependency: This script requires `docker for mac` to be installed: https://docs.docker.com/docker-for-mac/install/
+
 # Required parameters:
 # @raycast.schemaVersion 1
-# @raycast.title Clean Docker
+# @raycast.title Clean
 # @raycast.mode compact
 # @raycast.packageName Docker
 
@@ -12,6 +14,13 @@
 
 # Documentation:
 # @raycast.description Script that cleans Docker images, volumes, and containers
+# @raycast.author Quentin Eude
+# @raycase.authorURL https://github.com/qeude
+
+if ! command -v docker &> /dev/null; then
+      echo "docker for mac is required (https://docs.docker.com/docker-for-mac/install/).";
+      exit 1;
+fi
 
 docker stop $(docker ps -a -q) && docker rm $(docker ps -a -q)
 docker image prune -a -f
